+++ src/lib.rs
@@ -30,6 +30,9 @@
     missing_docs,
     warnings
 )]
+#![cfg_attr(not(feature = "std"), no_std)]
+extern crate alloc;
+extern crate core;
 
 #[cfg(feature = "derive")]
 #[macro_use]
@@ -38,10 +41,11 @@ extern crate amplify_derive;
 pub use amplify_derive::{Wrapper, Display, AsAny, From, Getters, Error};
 
 #[cfg(feature = "serde")]
+#[macro_use]
 extern crate serde_crate as serde;
 
 pub extern crate amplify_num as num;
-#[cfg(feature = "std")]
+#[cfg(any(test, feature = "hex"))]
 pub use num::hex;
 #[cfg(feature = "stringly_conversions")]
 pub extern crate stringly_conversions;
@@ -64,6 +68,7 @@ mod dumb_default;
 mod io_error;
 #[cfg(feature = "c_raw")]
 mod raw;
+mod slice32;
 pub mod strategy;
 #[cfg(feature = "serde")]
 mod to_serde_string;
@@ -75,6 +80,7 @@ pub use crate::as_any::AsAny;
 pub use crate::bipolar::Bipolar;
 pub use crate::strategy::Holder;
 pub use crate::wrapper::Wrapper;
+pub use crate::slice32::Slice32;
 pub use crate::dumb_default::DumbDefault;
 #[cfg(feature = "serde")]
 pub use crate::to_serde_string::{ToYamlString, ToJsonString, ToTomlString};